fix: change download

+++ hello_cangjie_eur.spec
@@ -12,14 +12,6 @@ A demo for Cangjie and Eur.
 
 %global debug_package %{nil}
 
-# 检查文件是否存在，并设置一个宏
-if [ ! -f %{_builddir}/Cangjie-0.53.13-linux_x64.tar.gz ]; then
-   %global file_not_exists 1
-else
-   %global file_not_exists 0
-fi
-
-
 # 根据环境设置下载地址
 %ifarch x86_64
 # 这里的指令仅在 x86_64 架构下运行
@@ -38,9 +30,10 @@ fi
 %build
 cd %{_builddir}
 
-%if %{file_not_exists}
-  wget -O Cangjie-0.53.13-linux.tar.gz %{download_url}
-%endif
+# 检查文件是否存在，并设置一个宏
+if [ ! -f %{_builddir}/Cangjie-0.53.13-linux.tar.gz ]
+   wget -O Cangjie-0.53.13-linux.tar.gz %{download_url}
+fi
 
 tar xvf Cangjie-0.53.13-linux.tar.gz
